+++ src/index.ts
@@ -27,6 +27,7 @@ export { default as Switch } from '@/package/components/switch.vue';
 export { default as Popover } from '@/package/components/popover.vue';
 export { default as Modal } from '@/package/components/modal.vue';
 export { default as Drawer } from '@/package/components/drawer.vue';
+export { default as Dropdown } from '@/package/components/dropdown.vue';
 export { default as Tooltip } from '@/package/components/tooltip.vue';
 export { default as Skeleton } from '@/package/components/skeleton.vue';
 export { default as Progress } from '@/package/components/progress.vue';